ui/requirements.txt
gradio>=4.0.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
PyYAML>=6.0
pydantic>=2.0.0